// src/models/apiTaxComplianceTvrPeriodListWorkflowStatus.ts
/**
 * Generated by orval
 * Do not edit manually.
 */

export type ApiTaxComplianceTvrPeriodListWorkflowStatus =
  (typeof ApiTaxComplianceTvrPeriodListWorkflowStatus)[keyof typeof ApiTaxComplianceTvrPeriodListWorkflowStatus]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApiTaxComplianceTvrPeriodListWorkflowStatus = {
  DRAFT: 'DRAFT',
  FUNDING_RECEIVED: 'FUNDING_RECEIVED',
  PREPARED: 'PREPARED',
  PUBLISHED: 'PUBLISHED',
  REVIEW_COMMENTS: 'REVIEW_COMMENTS',
} as const